Fix a panic case in the postgres_fdw test.
It happens in the below query:
 explain (verbose, costs off)
     select foo.f1, loct1.f1 from foo join loct1 on (foo.f1 = loct1.f1) order by foo.f2 offset 10 limit 10;
	  select foo.f1, loct1.f1 from foo join loct1 on (foo.f1 = loct1.f1) order by foo.f2 offset 10 limit 10;

The reason is that set_append_path_locus()->cdbpath_create_motion_path()
returns NULL (i.e. can not create a motion to reach the target locus
requirement) but set_append_path_locus() does not handle this bad case and
panic due to this (i.e. when accessing subpath->sameslice_relids while subpath
is NULL). This happens because in the above query, foo has a child with foreign
table (and a local heap table child additionally) and the append node tries to
create a parameterized path (one is an index scan path and another is the
foreign scan path; the target locus is Entry).  This happens when
use_remote_estimate is set on the foreign table. Per postgresGetForeignPaths()
code only when this option is set, parameterized path for foreign scan is
considered.

Fixing this by returning NULL path for the case that set_append_path_locus()
returns false.  Callers of them should take care of the NULL returning cases.
We did not modify all the scenarios, but will create a seperate story to track
that though that is not that urgent or serious since we previously did not see
such scenario.

Co-authored-by: Huiliang.liu <lhuiliang@vmware.com>